refactor: move .tvmaze/.imdb writers into proper plugins
