+++ etc/passwd
@@ -1,2 +1,1 @@
 goftpd:$c8aa2099$89be575337e36892c6d7f4181cad175d685162ad:0:5:0:/site:/bin/false
-
